tests/test_data/tf_bert_en_uncased/requirements.txt
tensorflow_text==2.3.0
tf-models-official==2.3.0
tensorflow_hub
uvicorn[standard]
fastapi
